requirements.txt
alembic==1.5.4
autopep8==1.5.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.4
Flask-Cors==4.0.0
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
passlib==1.7.4
postgres==4.0
psycopg==3.1.10
psycopg2==2.9.7
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pycodestyle==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typing_extensions==4.7.1
urllib3==1.26.4
Werkzeug==1.0.1
